LPS-61188 Simplify - PortalSettingsBaseDynamicInclude no longer needed
